requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
eodhd==1.0.32
fonttools==4.60.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
patsy==1.0.1
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rich==14.1.0
scipy==1.16.2
six==1.17.0
SQLAlchemy==2.0.43
statsmodels==0.14.5
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
